chore: moved logic to util
